fixed segfault on single quote

+++ hw2/structures.h
@@ -10,7 +10,7 @@ typedef struct tokenitem {
    char *text;     /* the actual string (lexeme) matched */
    int lineno;     /* the line number on which the token occurs */
    char *filename; /* the source file in which the token occurs */
-   int ival;       /* if you had an integer constant, store its value here */
+   double ival;       /* if you had an integer constant, store its value here */
    char *sval;      /* if you had a string constant, malloc space and store */
 } token ;        /* the string (less quotes and after escapes) here */
